move polling_plan to plan_stubs, keep in __all__ of plans to retain backwards compatibility
