+++ src/ibex_bluesky_core/plans/__init__.py
@@ -5,8 +5,7 @@ from typing import TYPE_CHECKING, Any
 
 import bluesky.plans as bp
 from bluesky import plan_stubs as bps
-from bluesky.plan_stubs import trigger_and_read
-from bluesky.protocols import NamedMovable, Readable
+from bluesky.protocols import NamedMovable
 from bluesky.utils import Msg
 from ophyd_async.plan_stubs import ensure_connected
 
@@ -14,7 +13,8 @@ from ibex_bluesky_core.callbacks import ISISCallbacks
 from ibex_bluesky_core.devices.block import BlockWriteConfig, block_rw
 from ibex_bluesky_core.devices.simpledae import monitor_normalising_dae
 from ibex_bluesky_core.fitting import FitMethod
-from ibex_bluesky_core.utils import NamedReadableAndMovable, centred_pixel
+from ibex_bluesky_core.plan_stubs import polling_plan
+from ibex_bluesky_core.utils import centred_pixel
 
 if TYPE_CHECKING:
     from ibex_bluesky_core.devices.simpledae import SimpleDae
@@ -284,47 +284,3 @@ def motor_adaptive_scan(  # noqa: PLR0913
             rel=rel,
         )
     )
-
-
-def polling_plan(
-    motor: NamedReadableAndMovable, readable: Readable[Any], destination: float
-) -> Generator[Msg, None, None]:
-    """Move to a destination but drop updates from readable if motor position has not changed.
-
-    Note - this does not start a run, this should be done with a run_decorator or similar in an
-    outer plan which calls this plan.
-
-    Args:
-        motor: the motor to move.
-        readable: the readable to read updates from, but drop if motor has not moved.
-        destination: the destination position.
-
-    Returns:
-        None
-
-    If we just used bp.scan() with a readable that updates more frequently than a motor can
-    register that it has moved, we would have lots of updates with the same motor position,
-    which may not be helpful.
-
-    """
-    yield from bps.checkpoint()
-    yield from bps.create()
-    reading = yield from bps.read(motor)
-    yield from bps.read(readable)
-    yield from bps.save()
-
-    # start the ramp
-    status = yield from bps.abs_set(motor, destination, wait=False)
-    while not status.done:
-        yield from bps.create()
-        new_reading = yield from bps.read(motor)
-        yield from bps.read(readable)
-
-        if new_reading[motor.name]["value"] == reading[motor.name]["value"]:
-            yield from bps.drop()
-        else:
-            reading = new_reading
-            yield from bps.save()
-
-    # take a 'post' data point
-    yield from trigger_and_read([motor, readable])
